fix bug: 文件打开的方式错了 2333

+++ typedata.h
@@ -35,18 +35,18 @@ ListNode *new_list_node(); //新建一个链表节点
 /* 请求结构体 */
 typedef struct {
     char *read_cache;       //缓存读取内容
-    int read_cache_len;     //缓存的内容长度
+    long read_cache_len;     //缓存的内容长度
     char *header_dump;      //缓存请求头
     int header_dump_len;    //请求头的长度
     char method;            //请求方法
     char *request_path;     //请求路径
     char *tmp_file_path;    //临时文件
-    int body_len;          //请求的body长度
+    long body_len;          //请求的body长度
 } Req;
 
 /* 响应结构体 */
 typedef struct {
-    int response_cache_len;    //响应内容长度
+    long response_cache_len;    //响应内容长度
     char *response_path;        //响应文件路径
 } Resp;
